pyproject.toml
[project]
name = "simple-discord"
version = "0.1.0"
description = "A simple discord backend service"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic-settings>=2.12.0",
    "aioboto3>=15.5.0",
    "aiofiles>=25.1.0",
    "python-ulid>=3.1.0",
    "typing-inspection>=0.4.2",
    "annotated-doc>=0.0.4",
]

[tool.setuptools.packages.find]
where = ["src"]